+++ packages/cca-scripts/index.js
@@ -1,0 +1,7 @@
+const util = require('./scripts/util');
+
+module.exports = {
+	collectLambdas() {
+		return util.lambda.all();
+	}
+};