+++ packages/cca-scripts/scripts/get-client-config.js
@@ -3,7 +3,11 @@ const util = require('./util');
 util.cloudformation.getStackOutputs().then(outputs => {
 	let config = {
 		baseUrl: outputs.apiUrl,
-		apiKey: outputs.apiKey
+		apiKey: outputs.apiKey,
+		userPoolClientId: outputs.userPoolClientId,
+		userPoolId: outputs.userPoolId,
+		identityPoolId: outputs.identityPoolId,
+		region: util.app.region
 	};
 
 	console.log(JSON.stringify(config, null, 2));