+++ scripts/test.sh
@@ -1,2 +1,3 @@
 rm -rf my-app/
 create-cloud-api my-app
+cd my-app && npm run template -- cognito && npm run deploy