+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## WIP
+
+#### `create-cloud-api@1.0.4`
+
+* Updated *README.md*
+
+#### `cca-scripts@1.3.0`
+
+* Updated documentation
+* Cognito support
+* Lambda *package.json* `cca_scripts` section
+
 ## December 28, 2017
 
 #### `cca-scripts@1.2.4`
@@ -8,7 +20,7 @@
 
 #### `create-cloud-api@1.0.3`
 
-* Added CHANGELOG.md
+* Added *CHANGELOG.md*
 
 #### `cca-scripts@1.2.3`